Add DCAD Data Files
This script is used to join together the columns that we actually need for this project from the DCAD data files. The DCAD appraisal values are stored in folders that include many different 'tables' (csv files) so we need to merge certain columns in order to end up with the data (features) we want to use to train the model (in this case I pulled only the columns that will match the data from CCAD and TAD).